Adding of PicDimensionScan
This is was previously PicAverageDim. It's intension is the be able to
scan a directory and either (based on user input) return average pic
dimensions or place any image over a certain width in a different
folder. Generally used to prep for manually cropping side images that
did not crop correctly.